优化服务路由，服务及日志监控，对象数组序列化问题，多用户同时运行数据隔离

+++ apics/src/main/java/cn/jmicro/common/Constants.java
@@ -27,7 +27,7 @@ public interface Constants {
 	
 	public static final byte USE_SYSTEM_CLIENT_ID = -2;
 	
-	public static final byte NO_CLIENT_ID = -1;
+	public static final int NO_CLIENT_ID = -1;
 	
 	public static final byte LIMIT_TYPE_LOCAL = 1;
